# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "ATOS", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/analytics/cardio": {
      "get": {
        "tags": ["analytics"],
        "summary": "Read Cardio Analytics",
        "operationId": "analytics-read_cardio_analytics",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date" },
                { "type": "null" }
              ],
              "title": "From Date"
            }
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date" },
                { "type": "null" }
              ],
              "title": "To Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardioAnalyticsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/dashboard": {
      "get": {
        "tags": ["analytics"],
        "summary": "Read Dashboard",
        "operationId": "analytics-read_dashboard",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "week_start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Monday that starts the requested week",
              "title": "Week Start"
            },
            "description": "Monday that starts the requested week"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DashboardPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/cardio/monthly": {
      "get": {
        "tags": ["analytics"],
        "summary": "Read Cardio Monthly",
        "description": "Serie mensual de volumen cardio (distancia, desnivel y sesiones).",
        "operationId": "analytics-read_cardio_monthly",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "months",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 12, "title": "Months" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "object", "additionalProperties": true },
                  "title": "Response Analytics-Read Cardio Monthly"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/cardio/best-paces": {
      "get": {
        "tags": ["analytics"],
        "summary": "Read Cardio Best Paces",
        "description": "Mejor ritmo promedio por distancia objetivo (5k, 10k, 15k, 21.1k).",
        "operationId": "analytics-read_cardio_best_paces",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": { "additionalProperties": true, "type": "object" },
                  "type": "array",
                  "title": "Response Analytics-Read Cardio Best Paces"
                }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
    "/api/v1/analytics/cardio/hr-zones": {
      "get": {
        "tags": ["analytics"],
        "summary": "Read Cardio Hr Zones",
        "description": "Tiempo por zona de FC por semana (\u00faltimas N semanas).",
        "operationId": "analytics-read_cardio_hr_zones",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "weeks",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 12, "title": "Weeks" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HrZonesPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/cardio/hr-trend": {
      "get": {
        "tags": ["analytics"],
        "summary": "Read Cardio Hr Trend",
        "description": "FC media y m\u00e1xima por semana.",
        "operationId": "analytics-read_cardio_hr_trend",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "weeks",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 12, "title": "Weeks" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HrTrendPublic" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/activities/summary": {
      "get": {
        "tags": ["analytics"],
        "summary": "Read Activities Summary",
        "description": "Resumen mensual de actividades cardio de Strava por tipo.\n\n- by_type agrupa por sport_type (Run, Ride, Walk, ...).\n- coverage es la fracci\u00f3n de actividades del per\u00edodo con sport_type no nulo.\n- monthly.target_activities es una PROYECCI\u00d3N de la meta semanal existente\n  (WeeklyGoal.target_cardio_minutes o target_km) multiplicada por ~4.33\n  semanas/mes. No es una meta mensual real: si no hay WeeklyGoal, se\n  devuelve null.\n- calendar incluye todos los d\u00edas del mes (con types vac\u00edo si no hubo\n  actividad).",
        "operationId": "analytics-read_activities_summary",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "description": "Mes en formato YYYY-MM",
              "title": "Month"
            },
            "description": "Mes en formato YYYY-MM"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivitySummaryPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivitySummaryPublic": {
        "properties": {
          "month": { "type": "string", "title": "Month" },
          "period_start": { "type": "string", "title": "Period Start" },
          "period_end": { "type": "string", "title": "Period End" },
          "total_activities": {
            "type": "integer",
            "title": "Total Activities"
          },
          "coverage": { "type": "number", "title": "Coverage" },
          "by_type": {
            "items": { "$ref": "#/components/schemas/ActivityTypeSummary" },
            "type": "array",
            "title": "By Type"
          },
          "monthly": { "$ref": "#/components/schemas/ActivitySummaryMonthly" },
          "calendar": {
            "items": {
              "$ref": "#/components/schemas/ActivitySummaryCalendarDay"
            },
            "type": "array",
            "title": "Calendar"
          }
        },
        "type": "object",
        "required": [
          "month",
          "period_start",
          "period_end",
          "total_activities",
          "coverage",
          "by_type",
          "monthly",
          "calendar"
        ],
        "title": "ActivitySummaryPublic"
      },
...
      "CardioAnalyticsPublic": {
        "properties": {
          "from_date": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "null" }
            ],
            "title": "From Date"
          },
          "to_date": {
            "anyOf": [
              { "type": "string", "format": "date" },
              { "type": "null" }
            ],
            "title": "To Date"
          },
          "distance_meters": { "type": "number", "title": "Distance Meters" },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          },
          "average_pace_seconds_per_km": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Average Pace Seconds Per Km"
          },
          "sessions": { "type": "integer", "title": "Sessions" }
        },
        "type": "object",
        "required": [
          "from_date",
          "to_date",
          "distance_meters",
          "duration_seconds",
          "average_pace_seconds_per_km",
          "sessions"
        ],
        "title": "CardioAnalyticsPublic"
      },
...
      "DashboardDay": {
        "properties": {
          "date": { "type": "string", "format": "date", "title": "Date" },
          "activities": {
            "items": { "$ref": "#/components/schemas/DashboardActivity" },
            "type": "array",
            "title": "Activities"
          }
        },
        "type": "object",
        "required": ["date", "activities"],
        "title": "DashboardDay"
      },
      "DashboardKPIs": {
        "properties": {
          "cardio_distance_meters": {
            "type": "number",
            "title": "Cardio Distance Meters"
          },
          "cardio_duration_seconds": {
            "type": "integer",
            "title": "Cardio Duration Seconds"
          },
          "cardio_avg_pace_seconds_per_km": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Cardio Avg Pace Seconds Per Km"
          },
          "strength_volume_kg": {
            "type": "number",
            "title": "Strength Volume Kg"
          },
          "strength_total_sets": {
            "type": "integer",
            "title": "Strength Total Sets"
          },
          "sessions": { "type": "integer", "title": "Sessions" },
          "active_days": { "type": "integer", "title": "Active Days" }
        },
        "type": "object",
        "required": [
          "cardio_distance_meters",
          "cardio_duration_seconds",
          "cardio_avg_pace_seconds_per_km",
          "strength_volume_kg",
          "strength_total_sets",
          "sessions",
          "active_days"
        ],
        "title": "DashboardKPIs"
      },
      "DashboardPublic": {
        "properties": {
          "week_start": {
            "type": "string",
            "format": "date",
            "title": "Week Start"
          },
          "week_end": {
            "type": "string",
            "format": "date",
            "title": "Week End"
          },
          "kpis": { "$ref": "#/components/schemas/DashboardKPIs" },
          "timeline": {
            "items": { "$ref": "#/components/schemas/DashboardDay" },
            "type": "array",
            "title": "Timeline"
          },
          "upcoming_race": {
            "anyOf": [
              { "$ref": "#/components/schemas/UpcomingRacePublic" },
              { "type": "null" }
            ]
          },
          "previous_kpis": {
            "anyOf": [
              { "$ref": "#/components/schemas/DashboardKPIs" },
              { "type": "null" }
            ]
          },
          "sync_state": {
            "items": { "$ref": "#/components/schemas/SyncStateEntry" },
            "type": "array",
            "title": "Sync State"
          }
        },
        "type": "object",
        "required": ["week_start", "week_end", "kpis", "timeline"],
        "title": "DashboardPublic"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HrTrendPoint": {
        "properties": {
          "week": { "type": "string", "title": "Week" },
          "avg_hr": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Avg Hr"
          },
          "max_hr": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Max Hr"
          }
        },
        "type": "object",
        "required": ["week"],
        "title": "HrTrendPoint"
      },
      "HrTrendPublic": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/HrTrendPoint" },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "title": "HrTrendPublic"
      },
      "HrZonesPublic": {
        "properties": {
          "weeks": {
            "items": { "additionalProperties": true, "type": "object" },
            "type": "array",
            "title": "Weeks"
          }
        },
        "type": "object",
        "title": "HrZonesPublic"
      },
...
      "UpcomingRacePublic": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "event_name": { "type": "string", "title": "Event Name" },
          "date": { "type": "string", "format": "date-time", "title": "Date" },
          "location": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Location"
          },
          "distance_km": { "type": "number", "title": "Distance Km" }
        },
        "type": "object",
        "required": ["id", "event_name", "date", "location", "distance_km"],
        "title": "UpcomingRacePublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" }
        }
      }
    }
  }
}